find-newsee-store/requirements.txt
dify_plugin>=0.1.0,<0.2.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
jieba>=0.42.1
python-Levenshtein>=0.21.1
